pyproject.toml
[tool.poetry]
name = "habits"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
redis = "^5.0.6"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
black = "^24.4.2"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
coverage = "^7.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
